platforms/python/research/nexural-research/requirements.txt
pandas>=2.2
numpy>=2.0
pyarrow>=16.0
polars>=1.0
duckdb>=1.0
matplotlib>=3.8
plotly>=5.22
scipy>=1.13
statsmodels>=0.14
scikit-learn>=1.5
pydantic>=2.7
pyyaml>=6.0
rich>=13.7
openpyxl>=3.1
jupyterlab>=4.2
ipykernel>=6.29
optuna>=3.6
joblib>=1.4
orjson>=3.10
watchdog>=4.0
fastapi>=0.111
uvicorn[standard]>=0.30
